Refactor to allow use of DeltaTimeModifier instead of an input field Changes to how the default scene was constructed has changed now that it needs the delta time, with a side effect being that because initialiser lists aren't used the memory pool need to be initialised manually to ensure its destructed after scenes are so that scene objects aren't incorrectly destructed twice.

+++ Source/Core/Settings.cpp
@@ -17,7 +17,7 @@ namespace Engine
 		return TargetFrameRate;
 	}
 
-	double Settings::GetTargetFrameTime() const
+	float Settings::GetTargetFrameTime() const
 	{
 		return TargetFrameTime;
 	}
